Remove CreatedOn output from Program.cs
The `Console.WriteLine` statement displaying the `CreatedOn`
property of the `Car` object was removed from the `Program`
class in `Program.cs`. This change simplifies the output
by omitting the creation timestamp or date of the car
instance. Other functionality, including the display of
`Brand`, `Model`, and `Wheels` properties, as well as the
`Drive` method and indexer usage, remains unaffected.